small refactor of benchmarks
