+++ crates/benches/bench_results.txt
@@ -10,3 +10,15 @@
 {"name":"benchmark 1500 transactions","sending_time":10000,"responses_span":565,"longest_confirmation_difference":0,"stats":{"estimated_tps":24.05867813693705,"relevant_blocks":[[64,59319],[1435,2987]]},"block_times":[2999,3000,2999,2999,3000,3000,2999,2999,2999,3000,2999,2999,3000,3000,3000,3000,3000,2999,2999,3000,3000,2999,2999,2999,3000,2999,3000,4955,59416,2983,3000,3000,2999,3680,59319,2987],"block_sizes":[0,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,0,0,0,64,1435,0,0,0,0,64,1435,0]}
 {"name":"benchmark 1750 transactions","sending_time":11689,"responses_span":649,"longest_confirmation_difference":0,"stats":{"estimated_tps":24.654985269033958,"relevant_blocks":[[95,67954],[1654,2985]]},"block_times":[3000,3000,2998,3000,3000,3000,2999,3000,2999,3000,2999,2999,3000,3000,2998,3021,3001,3000,2999,3000,3000,3000,2999,3000,2999,2999,3000,3000,5434,69512,2984,3000,3002,2998,6554,67954,2985],"block_sizes":[0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,0,0,0,0,64,1685,0,0,0,0,95,1654,0]}
 {"name":"benchmark 2000 transactions","sending_time":13458,"responses_span":818,"longest_confirmation_difference":0,"stats":{"estimated_tps":26.878756504551504,"relevant_blocks":[[60,71379],[1939,2992]]},"block_times":[3000,2999,3001,2998,2999,3001,2999,3000,2999,2999,3000,2999,3000,3020,3001,3000,2999,2999,3000,3000,2999,3000,3000,2999,3000,3000,3000,3000,2999,3000,3000,3314,78084,2991,3000,2999,3000,2999,3788,71379,2992,2999],"block_sizes":[0,0,1,1,1,0,2,1,1,1,1,1,1,0,1,1,0,1,1,0,1,0,0,0,0,0,0,0,0,0,0,0,69,1930,0,0,0,0,0,60,1939,0,0]}
+{"name":"benchmark 2 transactions","sending_time":29,"responses_span":0,"longest_confirmation_difference":0,"stats":{"estimated_tps":0.33338889814969164,"relevant_blocks":[[1,3001],[1,2998]]},"block_times":[2998,3000,2999,3000,3000,2999,3000,2999,3000,2999,3002,2997,3000,2999,3000,2999,3001,2998],"block_sizes":[2,1,0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0]}
+{"name":"benchmark 100 transactions","sending_time":740,"responses_span":51,"longest_confirmation_difference":0,"stats":{"estimated_tps":16.5,"relevant_blocks":[[64,3001],[35,2999]]},"block_times":[2999,2999,3000,3000,2999,3000,2999,3000,2999,3000,2999,3000,2999,2999,3001,2999,3000,2999,3034,3002,2999,4325,3001,2999],"block_sizes":[1,2,0,1,1,1,1,1,1,1,1,1,1,1,0,1,1,1,0,29,70,0,64,35,0]}
+{"name":"benchmark 200 transactions","sending_time":1429,"responses_span":118,"longest_confirmation_difference":0,"stats":{"estimated_tps":33.930093776641094,"relevant_blocks":[[199,8731],[199,2999]]},"block_times":[2999,3000,2999,3000,2999,3000,2999,3000,2999,2999,2999,3022,3001,3000,2999,3000,3000,3000,2999,3000,8343,8731,2999],"block_sizes":[1,1,1,0,2,1,1,1,1,1,1,0,1,1,0,1,1,0,1,0,0,199,199,0]}
+{"name":"benchmark 300 transactions","sending_time":2167,"responses_span":153,"longest_confirmation_difference":0,"stats":{"estimated_tps":23.076329397237014,"relevant_blocks":[[64,9959],[235,2998]]},"block_times":[2999,3000,2999,2999,2999,3000,2999,3000,3000,2999,2999,3000,2999,2999,3000,2999,3000,3000,2999,2999,4799,9722,2997,5514,9959,2998],"block_sizes":[1,1,1,0,1,2,1,1,1,1,1,0,1,1,1,1,1,0,1,0,0,64,235,0,64,235,0]}
+{"name":"benchmark 400 transactions","sending_time":2836,"responses_span":244,"longest_confirmation_difference":0,"stats":{"estimated_tps":23.458169204538773,"relevant_blocks":[[69,14012],[330,2997]]},"block_times":[3000,2999,3000,3002,2997,3115,3041,3020,3000,3000,3000,3004,2995,3000,3000,3000,2999,3000,2999,3000,2999,3570,13744,2997,3000,3300,14012,2997],"block_sizes":[1,1,1,0,1,1,1,1,2,1,1,0,1,1,1,1,1,0,1,0,0,0,64,335,0,0,69,330,0]}
+{"name":"benchmark 500 transactions","sending_time":3435,"responses_span":225,"longest_confirmation_difference":0,"stats":{"estimated_tps":32.77719390436153,"relevant_blocks":[[206,12227],[293,2997]]},"block_times":[2998,2999,3000,3000,2999,3000,2999,2999,3000,2999,2999,3000,2999,3000,3000,3053,3000,3000,3000,2999,3000,4280,18116,2997,2998,9264,12227,2997],"block_sizes":[1,1,1,0,1,1,1,1,1,1,2,0,1,1,1,1,0,1,1,0,0,0,64,435,0,0,206,293,0]}
+{"name":"benchmark 750 transactions","sending_time":5120,"responses_span":331,"longest_confirmation_difference":0,"stats":{"estimated_tps":23.78004254373432,"relevant_blocks":[[66,28504],[683,2993]]},"block_times":[2999,3000,2999,2999,2999,3000,3000,2999,3000,2999,3001,2998,3000,2999,3000,3000,2999,3000,2999,3000,3000,2999,3004,30787,2994,4045,28504,2993],"block_sizes":[0,2,1,0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,749,0,66,683,0]}
+{"name":"benchmark 1000 transactions","sending_time":6747,"responses_span":405,"longest_confirmation_difference":0,"stats":{"estimated_tps":28.199627392310735,"relevant_blocks":[[215,32434],[784,2992]]},"block_times":[2999,3000,3003,2995,3000,2999,2999,3001,2998,3000,3000,2999,3000,3000,3000,2999,3000,3000,2999,2999,3000,3000,2999,2999,3069,38348,2990,2999,3000,9696,32434,2992],"block_sizes":[0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,0,65,934,0,0,0,215,784,0]}
+{"name":"benchmark 1250 transactions","sending_time":8378,"responses_span":493,"longest_confirmation_difference":0,"stats":{"estimated_tps":32.189062419462914,"relevant_blocks":[[282,35806],[967,2996]]},"block_times":[2999,2999,2999,3000,2999,3000,3000,2999,3000,2999,3000,2999,3000,3021,3001,2999,2999,2999,3000,2999,3000,2999,3000,3000,2999,3000,3937,47710,2995,2999,2999,12545,35806,2996],"block_sizes":[0,1,1,1,0,2,1,1,1,1,1,1,0,1,1,0,1,1,0,1,0,0,0,0,0,0,0,65,1184,0,0,0,282,967,0]}
+{"name":"benchmark 1500 transactions","sending_time":10013,"responses_span":600,"longest_confirmation_difference":0,"stats":{"estimated_tps":26.335669987174757,"relevant_blocks":[[177,53933],[1322,2986]]},"block_times":[3000,2999,2999,3000,2999,3000,3000,2998,3000,2999,3071,3001,2999,3000,3000,2999,3008,3001,3000,2999,2999,2999,2999,3000,2999,2999,3000,6270,56902,2986,3000,3000,2999,8040,53933,2986],"block_sizes":[0,1,1,1,0,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,0,0,0,105,1394,0,0,0,0,177,1322,0]}
+{"name":"benchmark 1750 transactions","sending_time":11527,"responses_span":643,"longest_confirmation_difference":0,"stats":{"estimated_tps":26.508836278759585,"relevant_blocks":[[206,62993],[1543,2985]]},"block_times":[2999,3000,3000,2999,3000,3001,2998,3000,3000,3000,3000,3000,2999,3000,2999,3000,3000,3000,2999,3000,2999,3000,3000,2999,2999,3000,2999,3000,3378,70570,2985,2999,2999,2999,10670,62993,2985],"block_sizes":[0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,1,1,0,0,0,0,0,0,0,0,0,20,1729,0,0,0,0,206,1543,0]}
+{"name":"benchmark 2000 transactions","sending_time":13365,"responses_span":796,"longest_confirmation_difference":0,"stats":{"estimated_tps":24.413776257938448,"relevant_blocks":[[60,78899],[1939,2981]]},"block_times":[3000,3000,2999,3000,3143,3000,2999,3000,2999,3001,2999,3000,2998,3000,3000,2999,3000,3000,3015,3000,2999,3000,3000,2999,3000,2999,3000,2999,3000,2999,3964,78742,2981,2999,3000,2999,3000,3427,78899,2981],"block_sizes":[0,0,1,1,1,1,1,1,1,1,1,1,1,0,2,1,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,64,1935,0,0,0,0,0,60,1939,0]}
